Move middleware to auth package
