+++ src/main/java/org/openpodcastapi/opa/auth/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package org.openpodcastapi.opa.config;
+package org.openpodcastapi.opa.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
